Multithreading
Implemented multithreading for the program. This is required to have all three people working at once (that being the enterer, the reviewer, and the approver), It is also required for file reading. The only change to our personal class is that the code should be written in the run() method instead of the constructor. As for the group classes i am working on implementing file locks, other than that everything else will be the same. (also the workflow tester isnt done yet but i will fix that)