invalid lookups cause instability, removed tests involving them
it seems that if a program with an open file crash out it takes time for the file that was being worked on to go into a state where windows will say that it is no longer open. as such, invalid lookups can cause database writes to fail, prompting the removal of  invalid lookups. since the fid is all backend, there should not be any issue with invalid lookups.